+++ conf.lua
@@ -1,3 +1,10 @@
 function love.conf(t)
-  t.window.title = "ninepatch demo"
+  t.window.title = "quilt demo"
+
+  t.modules.audio = false
+  t.modules.joystick = false
+  t.modules.physics = false
+  t.modules.sound = false
+  t.modules.thread = false
+  t.modules.video = false
 end