+++ init.lua
@@ -1,11 +1,15 @@
-local BASE = (...):gsub("init$", "")
-local NinePatch = require(BASE .. ".NinePatch")
+local BASE = (...):gsub("init$", ""):gsub("([^%.])$", "%1%.")
 
-local ninepatch = {
-  _NAME = "ninepatch",
-  _DESCRIPTION = "NinePatch graphics for the LÖVE game framework",
-  _VERSION = "1.0.0",
-  _URL = "https://github.com/binaryfs/love-ninepatch",
+--- @type quilt.NinePatch
+local NinePatch = require(BASE .. "NinePatch")
+
+--- Provides 9-patch graphics for the LÖVE game framework.
+--- @class quilt
+local quilt = {
+  _NAME = "quilt",
+  _DESCRIPTION = "9-patch graphics for the LÖVE game framework",
+  _VERSION = "1.1.0",
+  _URL = "https://github.com/binaryfs/quilt",
   _LICENSE = [[
     MIT License
 
@@ -31,19 +35,7 @@ local ninepatch = {
   ]],
 }
 
---- @param texture love.Texture Texture that contains the 9-patch
---- @param x integer X position of the 9-patch on the texture
---- @param y integer Y position of the 9-patch on the texture
---- @param w integer Width of the 9-patch
---- @param h integer Height of the 9-patch
---- @param mt integer Height of the top patch row
---- @param mr integer Width of the right patch column
---- @param mb integer Height of the bottom patch row
---- @param ml integer Width of the left patch column
---- @return ninepatch.NinePatch
---- @nodiscard
-function ninepatch.newNinePatch(texture, x, y, w, h, mt, mr, mb, ml)
-  return NinePatch.new(texture, x, y, w, h, mt, mr, mb, ml)
-end
+quilt.NinePatch = NinePatch
+quilt.newNinePatch = NinePatch.new
 
-return ninepatch
+return quilt